cadatro de usuários administradores

+++ src/store/userAdmin.ts
@@ -1,6 +1,6 @@
 import { defineStore } from "pinia";
 
-export const useUserStore = defineStore("user", () => {
+export const useUserAdminStore = defineStore("userAdmin", () => {
   const { api } = useAxios();
 
   const user = ref<UserProps>();
@@ -8,39 +8,38 @@ export const useUserStore = defineStore("user", () => {
   const $single = computed(() => user.value);
   const $all = computed(() => users.value);
 
-  const index = async (inputQuery: string, profileType?: ProfileType) => {
+  const index = async (inputQuery: string) => {
     const config = {
       params: {
         inputQuery,
-        profileType,
       },
     };
 
-    const { data } = await api.get<UserProps[]>("/user", config);
+    const { data } = await api.get<UserProps[]>("/user-admin", config);
 
     users.value = data;
   };
 
   const show = async (id: number) => {
-    const { data } = await api.get<UserProps>(`/user/${id}`);
+    const { data } = await api.get<UserProps>(`/user-admin/${id}`);
 
     user.value = data;
   };
 
   const create = async (input: UserProps) => {
-    const { data } = await api.post<UserProps>("/user", input);
+    const { data } = await api.post<UserProps>("/user-admin", input);
 
     user.value = data;
   };
 
   const update = async (input: UserProps) => {
-    const { data } = await api.put<UserProps>("/user", input);
+    const { data } = await api.put<UserProps>("/user-admin", input);
 
     user.value = data;
   };
 
   const destroy = async (id: number) => {
-    await api.delete(`/user/${id}`);
+    await api.delete(`/user-admin/${id}`);
   };
 
   return { $single, $all, index, create, update, destroy, show };
